Improve inter-process message queues.
The old code would fail if the runtime libraries break a write or read into
multiple requests, which is legal in POSIX. Fixing this is somewhat complicated,
since the message queues use non-blocking reads. The updated code sends a message
length as a header, making it possible to block until the full message arrives
at the receiver.