Add HSSI emulation to ASE
By default, HSSI is emulated as a loopback interface. Each HSSI channel's TX is routed back to the channel's RX port. The default loopback plugin can be replaced with a traffic emulation library by pointing ASE_HSSI_PLUGIN_PATH to an external library on the make command line.

HSSI emulation is disabled by default. To enable it, add enable-hssi-sim to the afu-top-interface section of the AFU's json file:

      "afu-top-interface":
         {
            "class": "ofs_plat_afu",
            "enable-hssi-sim": 1
         }